+++ src/main/java/csc1035/project2/Students.java
@@ -31,7 +31,7 @@ public class Students {
     private String lastname;
 
     @OneToMany(mappedBy = "student", fetch = FetchType.EAGER)
-    private Set<StudentBooking> studentBookings;
+    private Set<StudentBooking> studentBookings = new HashSet<>();
 
     @OneToMany( mappedBy = "students" , fetch = FetchType.EAGER)
     private Set<Take> takes = new HashSet<>();
@@ -62,23 +62,14 @@ public class Students {
      * @param o The object for comparison
      * @return returns true if all attributes in both objects are the same, or if they have the same memory address
      */
-
     @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Students students = (Students) o;
-        return Objects.equals(id, students.id) &&
-                Objects.equals(firstname, students.firstname) &&
-                Objects.equals(lastname, students.lastname) &&
-                Objects.equals(studentBookings, students.studentBookings) &&
-                Objects.equals(takes, students.takes);
+    public boolean equals(Object o){
+        if (this==o) return true;
+        if (o == null || o.getClass() != Staff.class) return false;
+        Staff s = (Staff) o;
+        return this.id.equals(s.getId()) && this.firstname.equals(s.getFirstname()) && this.lastname.equals(s.getLastname());
     }
 
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, firstname, lastname, studentBookings, takes);
-    }
 
     public String getId() {
         return id;